packages/tools/ussy-reverseoracle/pyproject.toml
[project]
name = "ussy-reverseoracle"
version = "0.1.0"
description = "Migrated from reverseoracle"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ussy-core",
    "click>=8.0",
    "httpx>=0.24",
    "pyyaml>=6.0",
]

[project.scripts]
ussy-reverseoracle = "ussy_reverseoracle.cli:main"
reverseoracle = "ussy_reverseoracle.legacy:deprecated_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ussy_reverseoracle"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
